Fix compilation issue in params.hpp and some other minor details
- The file cpp/include/params.hpp should not rely on external loggers
or any boost library because it is used by the dynamic openfst-style libraries,
so it is best to keep it minimal.
I've added a few simple defines so that it can compile log messages with (e.g. any other tool
thas has the logger header previously included) or without (i.e. dynamic openfst-style libraries).

- Also, make clean now deletes obj (e.g. necessary in cpp/test).
- build-test.sh now uses make clean instead of deleting files directly.
- make runtests has been disabled.
- test.sh separates prints java unit-tests and cpp unit tests into separate files.