Fixes for NPLM
1) Hifst uses an implementation of lm::EnumerateVocab
to map grammar and language model number ids. This kenlm functionality
is not available for NPLM, hence was never getting to run
and the output score was garbage.

I plugged in a nasty temporary hack which consists of loading
twice the NN model, the first time using nplm to get access
to the vocabulary, and call our implementation of EnumerateVocab
interface. The right thing to do on the long run is to change
the nplm wrapper so that this class gets to run automatically.

Though the NN models get loaded twice, these are too small to be noticeable,
so its no show stopper.

2) NPLM returns logn scores through the wrapper, whilst other KENLM flavours return
log10 scores. There is now a workaround in applylm and other tools.

3) Working around a hifst peculiarity, a small hack in the state and the scores
for <s> to ensure we always match external nplm testNeuralLM scores, in the same
way we expect to always match srilm ngram scores exactly, which is useful
for debugging purposes.