Add google test 1.7.0 to the repository
Added as a zip file, installation procedure simply doesn't
download it.

The reason is that the version of google test 1.7.0 in googletest.com
had unsupported autotools installation, which was being used by ucam-smt,
and it has been removed in github (now requires cmake or should be integrated
directly in the code). This is the easiest fix for now.