Support openfst 1.5.4
Changes in openfst sparse power weight semiring required
a few minor changes in ucam-smt to get all tools to compile.

There is an issue with openfst 1.5.* in which not all fst* operations
work correctly with our semiring libraries. To be investigated.

Regression tests have been modified so that minimal dependency is required to verify
ucam-smt tool outputs (i.e. use fstprint to convert to tropical).